# refactor(rust): Dispatch new-streaming IPC source to updated multiscan