--- a/py-polars/tests/unit/io/test_multiscan.py
+++ b/py-polars/tests/unit/io/test_multiscan.py
@@ -348,8 +348,8 @@ def test_schema_mismatch_type_mismatch(
 @pytest.mark.parametrize(
     ("scan", "write", "ext"),
     [
-        (pl.scan_ipc, pl.DataFrame.write_ipc, "ipc"),
         # (pl.scan_parquet, pl.DataFrame.write_parquet, "parquet"), # TODO: _
+        # (pl.scan_ipc, pl.DataFrame.write_ipc, "ipc"), # TODO: _
         pytest.param(
             pl.scan_csv,
             pl.DataFrame.write_csv,