Also does some drive-by refactoring.

#### Benchmarks on 100x 1M IPC files (50 cols each):
Performance is essentially identical.

Tested on AWS `c6gn.8xlarge`. Files are on local disk.

* `head(10_000_000)` (first 10 files)
```python
# Before (pypi 1.26.0)
8.03s user 3.97s system 780% cpu 1.538 total
# After
8.07s user 4.30s system 826% cpu 1.498 total
```

* `tail(10_000_000)` (last 10 files)
```python
# Before (pypi 1.26.0)
8.11s user 4.00s system 773% cpu 1.564 total
# After
8.11s user 4.56s system 845% cpu 1.498 total
```