@nameexhaustion 2025-04-03
Set to draft as I need to do a refactor to address some perf regressions